requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
python-telegram-bot>=20.0
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
transformers>=4.30.0
torch>=2.0.0
spacy>=3.7.0
nltk>=3.8.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
selenium>=4.15.0
webdriver-manager>=4.0.0
python-docx>=0.8.11 